refactor: domain, view 패키지로 관련 클래스 이관. cars 클래스에서 view 객체 의존성 제거
