+++ src/main/java/racinggamewinner/domain/Main.java
@@ -1,7 +1,9 @@
-package racinggamewinner;
+package racinggamewinner.domain;
 
 import racinggamewinner.random.RacingCarRandomGenerator;
 import racinggamewinner.random.RandomGenerator;
+import racinggamewinner.view.InputView;
+import racinggamewinner.view.ResultView;
 
 import java.util.Scanner;
 
@@ -9,13 +11,17 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         String carNames = InputView.nameOfCars(scanner);
-        int movingTryCount = InputView.movingTryCount(scanner);
+        int maxMovingCount = InputView.movingTryCount(scanner);
 
         Cars racingCars = Cars.registerCars(carNames);
 
         RandomGenerator randomGenerator = new RacingCarRandomGenerator();
         ResultView.printPreview();
-        racingCars.viewRacing(movingTryCount, randomGenerator);
+
+        for (int i = 0; i < maxMovingCount; i++) {
+            racingCars.raceOneRound(randomGenerator);
+            ResultView.printTotalCarPosition(racingCars.carNameAndPositions());
+        }
 
         ResultView.printTotalCarPosition(racingCars.carNameAndPositions());
         ResultView.printWinner(racingCars.findWinner());
